// src/generated-api/src/models/DomainCreateFavoritePublisherListRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Affiliate Backend API
 * API Server for Affiliate Backend Application
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DomainCreateFavoritePublisherListRequest
 */
export interface DomainCreateFavoritePublisherListRequest {
    /**
     * 
     * @type {string}
     * @memberof DomainCreateFavoritePublisherListRequest
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof DomainCreateFavoritePublisherListRequest
     */
    name: string;
}

/**
 * Check if a given object implements the DomainCreateFavoritePublisherListRequest interface.
 */
export function instanceOfDomainCreateFavoritePublisherListRequest(value: object): value is DomainCreateFavoritePublisherListRequest {
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function DomainCreateFavoritePublisherListRequestFromJSON(json: any): DomainCreateFavoritePublisherListRequest {
    return DomainCreateFavoritePublisherListRequestFromJSONTyped(json, false);
}

export function DomainCreateFavoritePublisherListRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): DomainCreateFavoritePublisherListRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'description': json['description'] == null ? undefined : json['description'],
        'name': json['name'],
    };
}

export function DomainCreateFavoritePublisherListRequestToJSON(json: any): DomainCreateFavoritePublisherListRequest {
    return DomainCreateFavoritePublisherListRequestToJSONTyped(json, false);
}

export function DomainCreateFavoritePublisherListRequestToJSONTyped(value?: DomainCreateFavoritePublisherListRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'description': value['description'],
        'name': value['name'],
    };
}